requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain_openai
langchain_chroma
python_dotenv
pypdf